added withLoader to authors
